requirements.txt
alembic==1.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cssmin==0.2.0
dataset==1.3.1
Flask==1.1.2
Flask-Assets==2.0
Flask-Cors==3.0.8
Flask-Login==0.5.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
tld==0.12.2
urllib3==1.25.9
waitress==1.4.3
webassets==2.0
Werkzeug==1.0.1
